Change JS file name
